Snce our backend is f32/i64/i8 (the ints are the default tyoes for f32 backends), we need to read and write i64 ints, not i32

+++ hw3/tests/hw3_tests.rs
@@ -27,7 +27,7 @@ json.dump(F.linear(x, w).flatten().tolist(), sys.stdout)
     ))
 }
 
-fn torch_cross_entropy(logits: Vec<f32>, shape: [usize; 2], targets: Vec<i32>) -> f32 {
+fn torch_cross_entropy(logits: Vec<f32>, shape: [usize; 2], targets: Vec<i64>) -> f32 {
     as_f32_vec(python_json(
         r#"
 import json
@@ -109,7 +109,7 @@ fn test_cross_entropy_loss() {
         &DEVICE,
     );
     let y: Tensor<MyAutodiffBackend, 1, Int> =
-        Tensor::from_data(TensorData::from([0i32, 2]), &DEVICE);
+        Tensor::from_data(TensorData::from([0i64, 2]), &DEVICE);
     let loss = cross_entropy_loss(logits, y);
     let loss_val: f32 = loss.into_scalar();
     let expected = torch_cross_entropy(vec![2.0, 1.0, 0.0, 0.0, 2.0, 1.0], [2, 3], vec![0, 2]);
@@ -123,7 +123,7 @@ fn test_cross_entropy_numerically_stable() {
         &DEVICE,
     );
     let y: Tensor<MyAutodiffBackend, 1, Int> =
-        Tensor::from_data(TensorData::from([1i32, 2]), &DEVICE);
+        Tensor::from_data(TensorData::from([1i64, 2]), &DEVICE);
     let loss = cross_entropy_loss(logits, y);
     let loss_val: f32 = loss.into_scalar();
     let expected = torch_cross_entropy(
@@ -148,7 +148,7 @@ fn test_sgd_step() {
     let x: Tensor<MyAutodiffBackend, 2> =
         Tensor::random([12, 4], Distribution::Normal(0.0, 1.0), &DEVICE);
     let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(
-        TensorData::from([0i32, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2]),
+        TensorData::from([0i64, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2]),
         &DEVICE,
     );
 
@@ -181,7 +181,7 @@ fn test_sgd_zero_grad() {
     let x: Tensor<MyAutodiffBackend, 2> =
         Tensor::random([4, 4], Distribution::Normal(0.0, 1.0), &DEVICE);
     let y: Tensor<MyAutodiffBackend, 1, Int> =
-        Tensor::from_data(TensorData::from([0i32, 1, 2, 0]), &DEVICE);
+        Tensor::from_data(TensorData::from([0i64, 1, 2, 0]), &DEVICE);
     let logits = layer.forward(x);
     let loss = cross_entropy_loss(logits, y);
     let _grads = loss.backward();
@@ -242,8 +242,8 @@ fn test_dataloader_content() {
         .unwrap();
     assert_eq!(xb0, x_expected);
 
-    let yb0: Vec<i32> = batches[0].1.clone().into_data().to_vec().unwrap();
-    let y_expected: Vec<i32> = y.clone().narrow(0, 0, 5).into_data().to_vec().unwrap();
+    let yb0: Vec<i64> = batches[0].1.clone().into_data().to_vec().unwrap();
+    let y_expected: Vec<i64> = y.clone().narrow(0, 0, 5).into_data().to_vec().unwrap();
     assert_eq!(yb0, y_expected);
 
     // Last batch: rows 10..12
@@ -289,8 +289,8 @@ fn test_dataloader_reiterable() {
             .to_vec()
             .unwrap();
         assert_eq!(v1, v2);
-        let l1: Vec<i32> = yb1.clone().into_data().to_vec().unwrap();
-        let l2: Vec<i32> = yb2.clone().into_data().to_vec().unwrap();
+        let l1: Vec<i64> = yb1.clone().into_data().to_vec().unwrap();
+        let l2: Vec<i64> = yb2.clone().into_data().to_vec().unwrap();
         assert_eq!(l1, l2);
     }
 }
@@ -304,7 +304,7 @@ fn test_epoch_eval() {
     let x: Tensor<MyAutodiffBackend, 2> =
         Tensor::random([11, 4], Distribution::Normal(0.0, 1.0), &DEVICE);
     let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(
-        TensorData::from([0i32, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1]),
+        TensorData::from([0i64, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1]),
         &DEVICE,
     );
 
@@ -347,7 +347,7 @@ fn test_epoch_train() {
     let x: Tensor<MyAutodiffBackend, 2> =
         Tensor::random([11, 4], Distribution::Normal(0.0, 1.0), &DEVICE);
     let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(
-        TensorData::from([0i32, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1]),
+        TensorData::from([0i64, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1]),
         &DEVICE,
     );
 
@@ -519,7 +519,7 @@ mod mnist {
         let magic = u32::from_be_bytes(data[0..4].try_into().unwrap());
         assert_eq!(magic, 2049);
         let n_labels = u32::from_be_bytes(data[4..8].try_into().unwrap()) as usize;
-        let labels: Vec<i32> = data[8..8 + n_labels].iter().map(|&l| l as i32).collect();
+        let labels: Vec<i64> = data[8..8 + n_labels].iter().map(|&l| l as i64).collect();
         Tensor::from_data(TensorData::new(labels, [n_labels]), &DEVICE)
     }
 
@@ -591,7 +591,7 @@ fn test_eval_two_layer_nn() {
 fn test_cross_entropy_loss_single_sample() {
     let logits: Tensor<MyAutodiffBackend, 2> =
         Tensor::from_data(TensorData::from([[1.0f32, 2.0, 3.0]]), &DEVICE);
-    let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(TensorData::from([2i32]), &DEVICE);
+    let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(TensorData::from([2i64]), &DEVICE);
     let loss: f32 = cross_entropy_loss(logits, y).into_scalar();
     let expected = torch_cross_entropy(vec![1.0, 2.0, 3.0], [1, 3], vec![2]);
     assert!(loss.is_finite());
@@ -605,7 +605,7 @@ fn test_cross_entropy_loss_perfect_prediction() {
         &DEVICE,
     );
     let y: Tensor<MyAutodiffBackend, 1, Int> =
-        Tensor::from_data(TensorData::from([0i32, 1]), &DEVICE);
+        Tensor::from_data(TensorData::from([0i64, 1]), &DEVICE);
     let loss: f32 = cross_entropy_loss(logits, y).into_scalar();
     assert!(loss.is_finite());
     assert!(
@@ -619,7 +619,7 @@ fn test_cross_entropy_loss_large_logits_stable() {
     let logits: Tensor<MyAutodiffBackend, 2> =
         Tensor::from_data(TensorData::from([[1e30f32, 0.0], [0.0, 1e30]]), &DEVICE);
     let y: Tensor<MyAutodiffBackend, 1, Int> =
-        Tensor::from_data(TensorData::from([0i32, 1]), &DEVICE);
+        Tensor::from_data(TensorData::from([0i64, 1]), &DEVICE);
     let loss: f32 = cross_entropy_loss(logits, y).into_scalar();
     assert!(
         loss.is_finite(),
@@ -650,7 +650,7 @@ fn test_sgd_moves_in_gradient_direction() {
 
     let x: Tensor<MyAutodiffBackend, 2> =
         Tensor::from_data(TensorData::from([[1.0f32, 0.0, 0.0, 0.0]]), &DEVICE);
-    let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(TensorData::from([0i32]), &DEVICE);
+    let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(TensorData::from([0i64]), &DEVICE);
 
     let logits = layer.forward(x);
     let loss = cross_entropy_loss(logits, y);
@@ -662,7 +662,7 @@ fn test_sgd_moves_in_gradient_direction() {
     let x2: Tensor<MyAutodiffBackend, 2> =
         Tensor::from_data(TensorData::from([[1.0f32, 0.0, 0.0, 0.0]]), &DEVICE);
     let y2: Tensor<MyAutodiffBackend, 1, Int> =
-        Tensor::from_data(TensorData::from([0i32]), &DEVICE);
+        Tensor::from_data(TensorData::from([0i64]), &DEVICE);
     let logits2 = layer.forward(x2);
     let loss2: f32 = cross_entropy_loss(logits2, y2).into_scalar();
     assert!(
@@ -675,7 +675,7 @@ fn test_sgd_moves_in_gradient_direction() {
 fn test_dataloader_single_sample() {
     let x: Tensor<MyAutodiffBackend, 2> =
         Tensor::from_data(TensorData::from([[1.0f32, 2.0]]), &DEVICE);
-    let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(TensorData::from([0i32]), &DEVICE);
+    let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(TensorData::from([0i64]), &DEVICE);
 
     let loader = DataLoader::new(x, y, 10);
     let batches: Vec<_> = loader.collect();
@@ -702,7 +702,7 @@ fn test_epoch_loss_decreases_with_training() {
         Tensor::random([20, 4], Distribution::Normal(0.0, 1.0), &DEVICE);
     let y: Tensor<MyAutodiffBackend, 1, Int> = Tensor::from_data(
         TensorData::from([
-            0i32, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1,
+            0i64, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1,
         ]),
         &DEVICE,
     );
